requirements.txt
arrow==1.3.0
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.4
numpy==1.26.2
pandas==2.1.3
Pint==0.22
pydantic==1.10.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
stravalib==1.5
types-python-dateutil==2.8.19.14
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
